% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cog}
\alias{cog}
\title{Sensor Locations}
\format{A data frame with 44 rows with the following variables:
\itemize{
  \item ID: An id for the sensor.
}}
\usage{
cog
}
\description{
A dataset containing time series cognostics for each sensor
}
\keyword{datasets}

