% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{launchApp}
\alias{launchApp}
\title{Launch shiny app for exploring pedestrian data}
\usage{
launchApp(prop = 0.01)
}
\arguments{
\item{prop}{what proportion of the raw data should be displayed?}
}
\description{
Using this function is not recommended (at least currently).
Visit the website for examples like this shiny app that don't
require shiny -- \url{http://cpsievert.github.io/pedestrians}
}
\examples{
\dontrun{
launchApp()
}
}

