% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pedestrians}
\alias{pedestrians}
\title{Pedestrian counts}
\format{A data frame with 831,639 rows with the following variables:
\itemize{
  \item DateTime: A date time when the count 
  \item ID: An id for the sensor location.
  \item Name: The name of the sensor.
  \item Counts: The number of pedestrians who walked by the sensor within the hour.
  \item Month: The month.
  \item Year: The year.
  \item Day: The day.
  \item Wday: The day of the week.
  \item Weekend: Is this a weekend?
  \item Hour: The hour.
}}
\usage{
pedestrians
}
\description{
A dataset containing pedestrian counts at various locations in the 
CBD area of Melbourne Australia between 2013 and 2016
}
\keyword{datasets}

