% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sensors}
\alias{sensors}
\title{Sensor Locations}
\format{A data frame with 44 rows with the following variables:
\itemize{
  \item ID: An id for the sensor location.
  \item Name: The name of the sensor.
  \item Description: A description of the sensor.
  \item Status: Is this sensor installed?
  \item Upload Date:
  \item Year Installed: 
  \item Location Type:
  \item Geometry:
  \item Latitude:
  \item Longitude:
}}
\usage{
sensors
}
\description{
A dataset containing information about each sensor.
}
\keyword{datasets}

