% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{launchApp}
\alias{launchApp}
\title{Launch shiny app for exploring pedestrian data}
\usage{
launchApp(prop = 0.01)
}
\arguments{
\item{prop}{what proportion of the raw data should be displayed?}
}
\description{
Launch shiny app for exploring pedestrian data
}
\examples{
\dontrun{
launchApp()
}
}

